# docs/openapi3.yaml
openapi: 3.0.0
info:
  description: Production-ready backend Endge-сервиса с эталонными
    RedPanda-обёртками и строгими архитектурными границами.
  title: Endge Service Backend API
  contact: {}
  version: 1.0.20
paths:
...
  "/api/v1/projects/{project_identity}/folders":
    get:
      security:
        - BearerAuth: []
      description: Возвращает неудаленные папки проекта для указанного entity type.
      tags:
        - folders
      summary: Список папок
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: components
          description: Entity type
          name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - components
              - converters
              - queries
              - data-views
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_folder.FoldersListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    post:
      security:
        - BearerAuth: []
      description: Создает папку внутри проекта и дерева указанного entity type.
      tags:
        - folders
      summary: Создать папку
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_http_v1_folder.CreateFolderRequest"
        description: Параметры папки
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_folder.FolderResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/folders/{folder_identity}":
    get:
      security:
        - BearerAuth: []
      description: Возвращает папку по project identity, folder identity и entity type.
      tags:
        - folders
      summary: Получить папку
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: shared-components
          description: Folder identity
          name: folder_identity
          in: path
          required: true
          schema:
            type: string
        - example: components
          description: Entity type
          name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - components
              - converters
              - queries
              - data-views
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_folder.FolderResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    delete:
      security:
        - BearerAuth: []
      description: Выполняет soft delete папки.
      tags:
        - folders
      summary: Удалить папку
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: shared-components
          description: Folder identity
          name: folder_identity
          in: path
          required: true
          schema:
            type: string
        - example: components
          description: Entity type
          name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - components
              - converters
              - queries
              - data-views
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    patch:
      security:
        - BearerAuth: []
      description: Обновляет папку и при необходимости перемещает ее по parent identity.
      tags:
        - folders
      summary: Обновить папку
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: shared-components
          description: Folder identity
          name: folder_identity
          in: path
          required: true
          schema:
            type: string
        - example: components
          description: Entity type
          name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - components
              - converters
              - queries
              - data-views
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_http_v1_folder.UpdateFolderRequest"
        description: Параметры обновления папки
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_folder.FolderResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/folders/{folder_identity}/hard":
    delete:
      security:
        - BearerAuth: []
      description: Выполняет hard delete папки, кроме system root folders.
      tags:
        - folders
      summary: Удалить папку физически
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: shared-components
          description: Folder identity
          name: folder_identity
          in: path
          required: true
          schema:
            type: string
        - example: components
          description: Entity type
          name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - components
              - converters
              - queries
              - data-views
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "409":
          description: Conflict
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/folders/{folder_identity}/restore":
    post:
      security:
        - BearerAuth: []
      description: Восстанавливает soft-deleted папку.
      tags:
        - folders
      summary: Восстановить папку
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: shared-components
          description: Folder identity
          name: folder_identity
          in: path
          required: true
          schema:
            type: string
        - example: components
          description: Entity type
          name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - components
              - converters
              - queries
              - data-views
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
...
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    internal_api_http_v1_folder.CreateFolderRequest:
      type: object
      required:
        - displayName
        - entityType
        - identity
      properties:
        description:
          type: string
          example: Reusable components
        displayName:
          type: string
          maxLength: 255
          minLength: 1
          example: Shared Components
        entityType:
          enum:
            - components
            - converters
            - queries
            - data-views
          allOf:
            - $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_domai\
                n_entities.FolderEntityType"
          example: components
        identity:
          type: string
          maxLength: 160
          minLength: 1
          example: shared-components
        meta:
          type: object
        parentIdentity:
          type: string
          example: root-components
    internal_api_http_v1_folder.FolderResponse:
      type: object
      properties:
        createdAt:
          type: string
          example: 2026-07-08T10:00:00Z
        deletedAt:
          type: string
          example: 2026-07-08T10:00:00Z
        description:
          type: string
          example: Reusable components
        displayName:
          type: string
          example: Shared Components
        entityType:
          enum:
            - components
            - converters
            - queries
            - data-views
          allOf:
            - $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_domai\
                n_entities.FolderEntityType"
          example: components
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000011
        identity:
          type: string
          example: shared-components
        isRoot:
          type: boolean
          example: false
        isSystem:
          type: boolean
          example: false
        meta:
          type: object
        parentIdentity:
          type: string
          example: root-components
        projectIdentity:
          type: string
          example: demo-project
        updatedAt:
          type: string
          example: 2026-07-08T10:00:00Z
    internal_api_http_v1_folder.FoldersListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/internal_api_http_v1_folder.FolderResponse"
    internal_api_http_v1_folder.UpdateFolderRequest:
      type: object
      required:
        - displayName
      properties:
        description:
          type: string
          example: Reusable components
        displayName:
          type: string
          maxLength: 255
          minLength: 1
          example: Shared Components
        meta:
          type: object
        parentIdentity:
          type: string
          example: root-components
...